# BitWaves.WebAPI/docs/swagger/bitwaves-v1.yaml
swagger: '2.0'
info:
  description: BitWaves 后端接口文档
  version: 1.0.0
  title: BitWaves Web API Docs
  contact:
    email: msrlancern@126.com
...
paths:
...
  /users:
    get:
      summary: 获取用户列表
      description: 获取用户列表
      tags:
        - users
      produces:
        - application/json
      parameters:
        - in: query
          name: page
          type: integer
          required: false
          description: 分页参数，页面编号
          minimum: 0
          default: 0
        - in: query
          name: itemsPerPage
          type: integer
          required: false
          description: 分页参数，每一页中显示的用户数量
          minimum: 1
          default: 20
        - in: query
          name: by
          type: string
          required: false
          description: 排序依据
          enum:
            - TotalSubmissions
            - TotalAccepted
            - TotalProblemsAttempted
            - TotalProblemsAccepted
          default: TotalProblemsAccepted
        - in: query
          name: descend
          type: boolean
          required: false
          description: 排序时是否应该按照给定的排序依据降序排序
          default: true
      responses:
        200:
          description: 操作成功完成
          headers:
            X-BitWaves-Count:
              type: integer
              description: 在未分页之前满足筛选条件的用户数量
          schema:
            type: array
            items:
              $ref: '#/definitions/UserListItem'
        400:
          description: 无效的查询参数
    post:
      summary: 创建一个新用户
      description: 创建一个新用户
      tags:
        - users
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/PostUserModel'
      responses:
        200:
          description: 操作成功完成
        400:
          description: 数据模型验证失败，响应的body中包含模型验证错误信息
        409:
          description: 给定的用户名已经存在
  /users/{username}:
    get:
      summary: 获取用户信息
      description: 获取用户信息
      tags:
        - users
      produces:
        - application/json
      parameters:
        - in: path
          name: username
          type: string
          required: true
          description: 用户名
        - in: query
          name: detailed
          type: boolean
          required: false
          description: 是否包含更加详细的信息，默认为 `false`
        - in: header
          name: Authorization
          type: string
          required: false
          description: 当 `detailed` 为 `true` 时，提供身份验证信息。内容格式应该为 `Jwt <jwt>`，注意 `Jwt` 与 `<jwt>` 之间的空格
      responses:
        200:
          description: 操作成功完成
          schema:
            $ref: '#/definitions/UserInfo'
        403:
          description: 参数 `detailed` 为 `true` 但提供的 JWT 无效
        404:
          description: 用户名未找到
    put:
      summary: 修改用户信息
      description: 修改用户信息
      tags:
        - users
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: username
          type: string
          required: true
          description: 用户名
        - in: header
          name: Authorization
          type: string
          required: true
          description: 提供身份验证信息。内容格式应该为 `Jwt <jwt>`，注意 `Jwt` 与 `<jwt>` 之间的空格
        - in: body
          name: data
          required: true
          description: 描述用户信息修改数据
          schema:
            $ref: '#/definitions/PutUserModel'
      responses:
        200:
          description: 操作成功完成
        400:
          description: 数据模型验证失败
        403:
          description: 用户不具备相应的权限
  /users/{username}/password:
    put:
      summary: 修改用户密码
      description: 修改用户密码
      tags:
        - users
      parameters:
        - in: path
          name: username
          type: string
          required: true
          description: 用户名
        - in: header
          name: Authorization
          required: true
          type: string
          description: 提供身份验证信息。内容格式应该为 `Jwt <jwt>`，注意 `Jwt` 与 `<jwt>` 之间的空格
        - in: body
          name: data
          required: true
          description: 修改密码数据
          schema:
            $ref: '#/definitions/PutUserPasswordModel'
      responses:
        200:
          description: 操作成功完成
        400:
          description: 请求数据无效
        403:
          description: 用户没有相应的权限
        404:
          description: 给定的用户未找到
...
definitions:
...
  PostUserModel:
    type: object
    required:
      - username
      - password
      - phone
    properties:
      username:
        description: 用户名
        minLength: 3
        maxLength: 64
        type: string
        example: Lancern
      password:
        description: 密码
        minLength: 6
        type: string
        example: hello_world
      phone:
        description: 手机号，由 11 个十进制数字组成
        type: string
        example: "13901234567"
  PutUserModel:
    type: object
    properties:
      nickname:
        type: string
        minLength: 1
        maxLength: 64
        description: 昵称。若不需要更改昵称，请不要提供该成员。
        example: "msr"
      phone:
        type: string
        description: 手机号码，包含 11 位数字。若不需要更改手机号，请不要提供该成员。
        example: "13701234567"
      email:
        type: string
        description: 电子邮箱。若不需要更改电子邮箱，请不要提供该成员。
        example: "msr@bit.edu.cn"
      school:
        type: string
        minLength: 1
        maxLength: 64
        description: 学校名称。若不需要更改学校名称，请不要提供该成员。
        example: "Beijing Institute of Technology"
      studentId:
        type: string
        minLength: 3
        maxLength: 30
        description: 学号，由 10 位数字组成。若不需要更改学号，请不要提供该成员。
        example: "1120160001"
      blogUrl:
        type: string
        description: 用户博客 URL。若不需要更改此项，请不要提供该成员。
        example: "https://lancern.github.io"
  PutUserPasswordModel:
    type: object
    required:
      - password
    properties:
      oldPassword:
        type: string
        description: 当用户不是管理员时，该字段表示旧密码；否则该字段将被忽略
        minLength: 6
        example: "bitwaves2019"
      password:
        type: string
        description: 新密码
        minLength: 6
        example: "bitwaves2020"
  UserListItem:
    type: object
    required:
      - username
      - nickname
      - joinTime
      - totalSubmissions
      - totalAccepted
      - totalProblemsAttempted
      - totalProblemsAccepted
    properties:
      username:
        type: string
        description: 用户名
        example: Lancern
      nickname:
        type: string
        description: 用户昵称
        example: MSR
      joinTime:
        type: string
        description: 用户加入 BITOJ 的时间，时区为 UTC 时区
        example: "2019-9-24T10:00:00"
      totalSubmissions:
        type: integer
        description: 用户总提交数量
        example: 10
      totalAccepted:
        type: integer
        description: 用户总 AC 提交数量
        example: 3
      totalProblemsAttempted:
        type: integer
        description: 用户总共尝试的题目数量
        example: 2
      totalProblemsAccepted:
        type: integer
        description: 用户总 AC 题目数量
        example: 1
  UserInfo:
    type: object
    required:
      - username
      - nickname
      - phone
      - email
      - school
      - studentId
      - blogUrl
      - joinTime
      - totalSubmissions
      - totalAccepted
      - totalProblemsAttempted
      - totalProblemsAccepted
      - isAdmin
    properties:
      username:
        type: string
        description: 用户名
        example: Lancern
      nickname:
        type: string
        description: 用户昵称
        example: msr
      phone:
        type: string
        description: 当 `detailed` 为 `true` 时，本字段包含用户的手机号
        example: "13901234567"
      email:
        type: string
        description: 用户邮箱
        example: msr@bit.edu.cn
      school:
        type: string
        description: 用户的学校
        example: Beijing Institute of Technology
      studentId:
        type: string
        description: 当 `detailed` 为 `true` 时，本字段包含用户的学号
        example: "1120160123"
      blogUrl:
        type: string
        description: 用户的博客 URL
        example: https://lancern.github.io
      joinTime:
        type: string
        description: 用户加入 BITOJ 的时间，时区为 UTC 时区
        example: "2019-9-24T10:00:00"
      totalSubmissions:
        type: integer
        description: 用户总提交数量
        example: 10
      totalAccepted:
        type: integer
        description: 用户总 AC 提交数量
        example: 3
      totalProblemsAttempted:
        type: integer
        description: 用户总共尝试的题目数量
        example: 2
      totalProblemsAccepted:
        type: integer
        description: 用户总 AC 题目数量
        example: 1
      isAdmin:
        type: boolean
        description: 用户是否为管理员
        example: true
...